integracao back-end front-end home - em implementacao - upload images 2

+++ server/comunideia/app/controllers/images_controller.rb
@@ -35,29 +35,48 @@ FlickRaw.shared_secret=SHARED_SECRET
 
   def upload_images
   	
-  	begin
-      login = flickr.test.login
-      
-      uploaded_io = params[:image]
-      image_name = uploaded_io.original_filename.blank? ? DateTime.now.strftime('%Y%m%d%H%M%S%L') : uploaded_io.original_filename
+    idea = Idea.find(params[:idea_id])
 
-      pic_path = Rails.root.join('public', 'uploads_tmp', image_name)
-      File.open(pic_path, 'wb') do |file|
-        file.write(uploaded_io.read)
+    i = 0
+    params[:images].each do |key, value|
+      if i < Idea::MAX_IMAGES
+
+        begin
+          login = flickr.test.login
+          
+          uploaded_io = value
+          image_name = uploaded_io.original_filename.blank? ? DateTime.now.strftime('%Y%m%d%H%M%S%L') : uploaded_io.original_filename
+
+          pic_path = Rails.root.join('public', 'uploads_tmp', image_name)
+          File.open(pic_path, 'wb') do |file|
+            file.write(uploaded_io.read)
+          end
+
+          # You need to be authentified to do that
+          photo_id = flickr.upload_photo pic_path, :title => image_name
+
+          info = flickr.photos.getInfo(:photo_id => photo_id)
+
+
+          File.delete(pic_path)
+
+          photo_url = FlickRaw.url_b(info).to_s
+          idea.img_pgs[i] = photo_url
+
+        rescue FlickRaw::OAuthClient::FailedResponse => e
+          flash[:error] = "Ocorreu um erro ao carregar sua imagem. Você pode enviar novamente?"
+        end
+
       end
 
-      # You need to be authentified to do that
-      flickr.upload_photo pic_path, :title => image_name
+      i += 1
 
-      File.delete(pic_path)
-
-      flash[:success] = "Imagem enviada."
-      redirect_to root_path #render 'idea-form'
-
-    rescue FlickRaw::OAuthClient::FailedResponse => e
-      flash[:error] = "Ocorreu um erro ao carregar sua imagem. Você pode enviar novamente?"
-      redirect_to root_path #render 'idea-form'
     end
+
+    idea.current_step = 0
+    idea.save
+      
+    redirect_to root_path #render 'idea-form'
   end
 
 end
